+++ src/MBH/Bundle/OnlineBundle/Resources/public/js/form.js
@@ -1,5 +1,16 @@
 /*global window, $, document */
 $(document).ready(function () {
     'use strict';
-    return;
+
+    // roomTypeChoices
+    (function (){
+        var roomTypes = $('#mbh_bundle_onlinebundle_form_type_roomTypes');
+        var choices = $('#mbh_bundle_onlinebundle_form_type_roomTypeChoices')
+            .closest('div.form-group');
+        var toggleChoices = function () {
+            choices.toggle(roomTypes.is(':checked'));
+        };
+        toggleChoices();
+        roomTypes.on('change switchChange.bootstrapSwitch', toggleChoices);
+    }());
 });