+++ src/MBH/Bundle/PackageBundle/Controller/TouristController.php
@@ -209,7 +209,7 @@ class TouristController extends Controller
 
             if ($entity->getBirthday() || $entity->getPhone() || $entity->getEmail()) {
                 $pieces = [];
-                !$entity->getBirthday() ?: $entity[] = $entity->getBirthday()->format('d.m.Y');
+                !$entity->getBirthday() ?: $pieces[] = $entity->getBirthday()->format('d.m.Y');
                 !$entity->getPhone() ?: $pieces[] = $entity->getPhone();
                 !$entity->getEmail() ?: $pieces[] = $entity->getEmail();
                 $text .= ' (' . implode(', ', $pieces) . ')';