+++ src/MBH/Bundle/PackageBundle/Resources/public/js/chessBoard/ChessBoardManager.ts
@@ -499,6 +499,8 @@ class ChessBoardManager {
                 return false;
             }
         });
+
+        return topOffset;
     }
 
     private getPackageLeftOffset(startDate) {