+++ src/MBH/Bundle/PackageBundle/Resources/public/js/package.js
@@ -302,6 +302,7 @@ var docReadyPackages = function () {
         "processing": true,
         "serverSide": true,
         "ordering": true,
+        "searchDelay": 2500,
         "ajax": {
             "url": Routing.generate('package_json'),
             "data": function (d) {