+++ src/MBH/Bundle/UserBundle/Security/AuthenticationSuccessHandler.php
@@ -3,6 +3,7 @@
 namespace MBH\Bundle\UserBundle\Security;
 
 use MBH\Bundle\ClientBundle\Service\Mbhs;
+use MBH\Bundle\ClientBundle\Service\Dashboard\Dashboard;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Http\Authentication\DefaultAuthenticationSuccessHandler;
@@ -19,11 +20,17 @@ class AuthenticationSuccessHandler extends DefaultAuthenticationSuccessHandler
     protected $mbhs;
 
     /**
+     * @var Dashboard
+     */
+    protected $dashboard;
+
+    /**
      * {@inheritdoc}
      */
-    public function __construct(HttpUtils $httpUtils, array $options, Mbhs $mbhs)
+    public function __construct(HttpUtils $httpUtils, array $options, Mbhs $mbhs, Dashboard $dashboard)
     {
         $this->mbhs = $mbhs;
+        $this->dashboard = $dashboard;
         parent::__construct($httpUtils, $options);
     }
 
@@ -32,7 +39,10 @@ class AuthenticationSuccessHandler extends DefaultAuthenticationSuccessHandler
      */
     public function onAuthenticationSuccess(Request $request, TokenInterface $token)
     {
+        $request->getSession()->set('mbh.justLogin', true);
+        $this->dashboard->notify();
         $this->mbhs->login($request->getClientIp());
+
         return parent::onAuthenticationSuccess($request, $token);
     }
 }